pyproject.toml

[project]
name = "snap2store"
version = "0.1.2"
description = "Auto add device frames to iOS/iPadOS screenshots for App Store."
authors = [
	{ name = "Lilong Zhang", email = "and2long@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = ["ios", "ipad", "screenshot", "app-store", "developer-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
	"Pillow>=9.0.0",
	"psd-tools>=1.9.0"
]

[project.urls]
"Homepage" = "https://github.com/and2long/snap2store"
"Bug Tracker" = "https://github.com/and2long/snap2store/issues"

[project.scripts]
snap2store = "snap2store.main:main"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["snap2store"]
include-package-data = true

[tool.setuptools.package-data]
snap2store = ["../psd/*.psd"]
